requirements.txt
torch==2.7.1+cu128
diffusers==0.34.0
transformers
accelerate
safetensors
pillow
xformers
protobuf
sentencepiece
